Add reusable AnimatedCounter component for stats
